backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.16
asyncpg==0.29.0
httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.7 
structlog==24.1.0
pydantic==1.10.15
python-dotenv==1.0.1
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 
jinja2==3.1.4 
celery==5.3.4 
redis==5.0.4 
openai==1.23.6
prometheus-fastapi-instrumentator==6.1.0 
aioredis==2.0.2 
pgvector==0.2.5 
aiolimiter==1.1.0
google-auth-oauthlib==1.2.0 
cryptography==42.0.5 
tiktoken==0.6.0 
forex-python==1.8 
jsonschema==4.22.0 